Refactor add-docker-files for easier reading/maintenance
