+++ configureWorkspace/configure.ts
@@ -3,6 +3,7 @@
  *  Licensed under the MIT License. See LICENSE.md in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import * as assert from 'assert';
 import * as fs from 'fs';
 import * as fse from 'fs-extra';
 import * as gradleParser from "gradle-to-js/lib/parser";
@@ -14,8 +15,15 @@ import { IActionContext, TelemetryProperties } from 'vscode-azureextensionui';
 import { ext } from '../extensionVariables';
 import { globAsync } from '../helpers/async';
 import { OS, Platform, promptForPort, quickPickOS, quickPickPlatform } from './config-utils';
+import { configureAspDotNetCore } from './configure_asp.net.core';
+import { configureGo } from './configure_go';
+import { configureJava } from './configure_java';
+import { configureDotNetCoreConsole } from './configure_net.core.console';
+import { configureNode } from './configure_node';
+import { configurePython } from './configure_python';
+import { configureRuby } from './configure_ruby';
 
-interface PackageInfo {
+export interface PackageInfo {
     npmStart: boolean; //has npm start
     cmd: string;
     fullCommand: string; //full command
@@ -45,419 +53,52 @@ export type ConfigureTelemetryProperties = {
     packageFileSubfolderDepth?: string; // 0 = project/etc file in root folder, 1 = in subfolder, 2 = in subfolder of subfolder, etc.
 };
 
+const generatorsByPlatform = new Map<Platform, {
+    genDockerFile: GeneratorFunction,
+    genDockerCompose: GeneratorFunction,
+    genDockerComposeDebug: GeneratorFunction
+}>();
+generatorsByPlatform.set('ASP.NET Core', configureAspDotNetCore);
+generatorsByPlatform.set('Go', configureGo);
+generatorsByPlatform.set('Java', configureJava);
+generatorsByPlatform.set('.NET Core Console', configureDotNetCoreConsole);
+generatorsByPlatform.set('Node.js', configureNode);
+generatorsByPlatform.set('Python', configurePython);
+generatorsByPlatform.set('Ruby', configureRuby);
+
+//////////////////////////////////
+
 // Note: serviceName includes the path of the service relative to the generated file, e.g. 'projectFolder1/myAspNetService'
 // tslint:disable-next-line:max-func-body-length
-function genDockerFile(serviceName: string, platform: string, os: string, port: string, { cmd, author, version, artifactName }: PackageInfo): string {
-    switch (platform.toLowerCase()) {
-        case 'node.js':
-
-            return `FROM node:8.9-alpine
-ENV NODE_ENV production
-WORKDIR /usr/src/app
-COPY ["package.json", "package-lock.json*", "npm-shrinkwrap.json*", "./"]
-RUN npm install --production --silent && mv node_modules ../
-COPY . .
-EXPOSE ${port}
-CMD ${cmd}`;
-
-        case 'go':
-
-            return `
-#build stage
-FROM golang:alpine AS builder
-WORKDIR /go/src/app
-COPY . .
-RUN apk add --no-cache git
-RUN go get -d -v ./...
-RUN go install -v ./...
-
-#final stage
-FROM alpine:latest
-RUN apk --no-cache add ca-certificates
-COPY --from=builder /go/bin/app /app
-ENTRYPOINT ./app
-LABEL Name=${serviceName} Version=${version}
-EXPOSE ${port}
-`;
-
-        case '.net core console':
-
-            if (os.toLowerCase() === 'windows') {
-                return `
-
-FROM microsoft/dotnet:2.0-runtime-nanoserver-1709 AS base
-WORKDIR /app
-
-FROM microsoft/dotnet:2.0-sdk-nanoserver-1709 AS build
-WORKDIR /src
-COPY ${serviceName}.csproj ${serviceName}/
-RUN dotnet restore ${serviceName}/${serviceName}.csproj
-WORKDIR /src/${serviceName}
-COPY . .
-RUN dotnet build ${serviceName}.csproj -c Release -o /app
-
-FROM build AS publish
-RUN dotnet publish ${serviceName}.csproj -c Release -o /app
-
-FROM base AS final
-WORKDIR /app
-COPY --from=publish /app .
-ENTRYPOINT ["dotnet", "${serviceName}.dll"]
-`;
-            } else {
-                return `
-FROM microsoft/dotnet:2.0-runtime AS base
-WORKDIR /app
-
-FROM microsoft/dotnet:2.0-sdk AS build
-WORKDIR /src
-COPY ${serviceName}.csproj ${serviceName}/
-RUN dotnet restore ${serviceName}/${serviceName}.csproj
-WORKDIR /src/${serviceName}
-COPY . .
-RUN dotnet build ${serviceName}.csproj -c Release -o /app
-
-FROM build AS publish
-RUN dotnet publish ${serviceName}.csproj -c Release -o /app
-
-FROM base AS final
-WORKDIR /app
-COPY --from=publish /app .
-ENTRYPOINT ["dotnet", "${serviceName}.dll"]
-`;
-            }
-
-        case 'asp.net core':
-
-            if (os.toLowerCase() === 'windows') {
-                return `
-FROM microsoft/aspnetcore:2.0-nanoserver-1709 AS base
-WORKDIR /app
-EXPOSE ${port}
-
-FROM microsoft/aspnetcore-build:2.0-nanoserver-1709 AS build
-WORKDIR /src
-COPY ${serviceName}.csproj ${serviceName}/
-RUN dotnet restore ${serviceName}/${serviceName}.csproj
-WORKDIR /src/${serviceName}
-COPY . .
-RUN dotnet build ${serviceName}.csproj -c Release -o /app
-
-FROM build AS publish
-RUN dotnet publish ${serviceName}.csproj -c Release -o /app
-
-FROM base AS final
-WORKDIR /app
-COPY --from=publish /app .
-ENTRYPOINT ["dotnet", "${serviceName}.dll"]
-`;
-            } else {
-                return `
-FROM microsoft/aspnetcore:2.0 AS base
-WORKDIR /app
-EXPOSE ${port}
-
-FROM microsoft/aspnetcore-build:2.0 AS build
-WORKDIR /src
-COPY ${serviceName}.csproj ${serviceName}/
-RUN dotnet restore ${serviceName}/${serviceName}.csproj
-WORKDIR /src/${serviceName}
-COPY . .
-RUN dotnet build ${serviceName}.csproj -c Release -o /app
-
-FROM build AS publish
-RUN dotnet publish ${serviceName}.csproj -c Release -o /app
-
-FROM base AS final
-WORKDIR /app
-COPY --from=publish /app .
-ENTRYPOINT ["dotnet", "${serviceName}.dll"]
-`;
-            }
-
-        case 'python':
-
-            return `
-# Python support can be specified down to the minor or micro version
-# (e.g. 3.6 or 3.6.3).
-# OS Support also exists for jessie & stretch (slim and full).
-# See https://hub.docker.com/r/library/python/ for all supported Python
-# tags from Docker Hub.
-FROM python:alpine
-
-# If you prefer miniconda:
-#FROM continuumio/miniconda3
-
-LABEL Name=${serviceName} Version=${version}
-EXPOSE ${port}
-
-WORKDIR /app
-ADD . /app
-
-# Using pip:
-RUN python3 -m pip install -r requirements.txt
-CMD ["python3", "-m", "${serviceName}"]
-
-# Using pipenv:
-#RUN python3 -m pip install pipenv
-#RUN pipenv install --ignore-pipfile
-#CMD ["pipenv", "run", "python3", "-m", "${serviceName}"]
-
-# Using miniconda (make sure to replace 'myenv' w/ your environment name):
-#RUN conda env create -f environment.yml
-#CMD /bin/bash -c "source activate myenv && python3 -m ${serviceName}"
-`;
-
-        case 'ruby':
-
-            return `
-FROM ruby:2.5-slim
-
-LABEL Name=${serviceName} Version=${version}
-EXPOSE ${port}
-
-# throw errors if Gemfile has been modified since Gemfile.lock
-RUN bundle config --global frozen 1
-
-WORKDIR /app
-COPY . /app
-
-COPY Gemfile Gemfile.lock ./
-RUN bundle install
-
-CMD ["ruby", "${serviceName}.rb"]
-`;
-
-        case 'java':
-            const artifact = artifactName ? artifactName : `${serviceName}.jar`;
-            return `
-FROM openjdk:8-jdk-alpine
-VOLUME /tmp
-ARG JAVA_OPTS
-ENV JAVA_OPTS=$JAVA_OPTS
-ADD ${artifact} ${serviceName}.jar
-EXPOSE ${port}
-ENTRYPOINT exec java $JAVA_OPTS -jar ${serviceName}.jar
-# For Spring-Boot project, use the entrypoint below to reduce Tomcat startup time.
-#ENTRYPOINT exec java $JAVA_OPTS -Djava.security.egd=file:/dev/./urandom -jar ${serviceName}.jar
-`;
-
-        default:
-
-            return `
-FROM docker/whalesay:latest
-LABEL Name=${serviceName} Version=${version}
-RUN apt-get -y update && apt-get install -y fortunes
-CMD /usr/games/fortune -a | cowsay
-`;
+function genDockerFile(serviceName: string, platform: Platform, os: OS | undefined, port: string | undefined, { cmd, author, version, artifactName }: Partial<PackageInfo>): string {
+    let generators = generatorsByPlatform.get(platform);
+    assert(generators, `Could not find dockerfile generator functions for "${platform}"`);
+    if (generators.genDockerFile) {
+        return generators.genDockerFile(serviceName, platform, os, port, { cmd, author, version, artifactName });
     }
 }
 
-function genDockerCompose(serviceName: string, platform: string, os: string, port: string): string {
-    switch (platform.toLowerCase()) {
-        case 'node.js':
-            return `version: '2.1'
+//////////////////////////////////
 
-services:
-  ${serviceName}:
-    image: ${serviceName}
-    build: .
-    environment:
-      NODE_ENV: production
-    ports:
-      - ${port}:${port}`;
-
-        case 'go':
-            return `version: '2.1'
-
-services:
-  ${serviceName}:
-    image: ${serviceName}
-    build: .
-    ports:
-      - ${port}:${port}`;
-
-        case '.net core console':
-            // we don't generate compose files for .net core
-            return `version: '2.1'
-
-services:
-  ${serviceName}:
-    image: ${serviceName}
-    build: .
-    ports:
-      - ${port}:${port}`;
-
-        case 'asp.net core':
-            // we don't generate compose files for .net core
-            return `version: '2.1'
-
-services:
-  ${serviceName}:
-    image: ${serviceName}
-    build: .
-    ports:
-      - ${port}:${port}`;
-
-        case 'python':
-            return `version: '2.1'
-
-services:
-  ${serviceName}:
-    image: ${serviceName}
-    build: .
-    ports:
-      - ${port}:${port}`;
-
-        case 'ruby':
-            return `version: '2.1'
-
-services:
-  ${serviceName}:
-    image: ${serviceName}
-    build: .
-    ports:
-      - ${port}:${port}`;
-
-        case 'java':
-            return `version: '2.1'
-
-services:
-  ${serviceName}:
-    image: ${serviceName}
-    build: .
-    ports:
-      - ${port}:${port}`;
-
-        default:
-            return `version: '2.1'
-
-services:
-  ${serviceName}:
-    image: ${serviceName}
-    build: .
-    ports:
-      - ${port}:${port}`;
+function genDockerCompose(serviceName: string, platform: Platform, os: OS | undefined, port: string): string {
+    let generators = generatorsByPlatform.get(platform);
+    assert(generators, `Could not find docker compose file generator function for "${platform}"`);
+    if (generators.genDockerCompose) {
+        return generators.genDockerCompose(serviceName, platform, os, port);
     }
 }
 
-// tslint:disable-next-line:max-func-body-length
-function genDockerComposeDebug(serviceName: string, platform: string, os: string, port: string, { fullCommand: cmd }: PackageInfo): string {
-    switch (platform.toLowerCase()) {
-        case 'node.js':
+//////////////////////////////////
 
-            const cmdArray: string[] = cmd.split(' ');
-            if (cmdArray[0].toLowerCase() === 'node') {
-                cmdArray.splice(1, 0, '--inspect=0.0.0.0:9229');
-                cmd = `command: ${cmdArray.join(' ')}`;
-            } else {
-                cmd = '## set your startup file here\n    command: node --inspect index.js';
-            }
-
-            return `version: '2.1'
-
-services:
-  ${serviceName}:
-    image: ${serviceName}
-    build: .
-    environment:
-      NODE_ENV: development
-    ports:
-      - ${port}:${port}
-      - 9229:9229
-    ${cmd}`;
-
-        case 'go':
-            return `version: '2.1'
-
-services:
-  ${serviceName}:
-    image: ${serviceName}
-    build:
-      context: .
-      dockerfile: Dockerfile
-    ports:
-        - ${port}:${port}
-`;
-
-        case '.net core console':
-            // we don't generate compose files for .net core
-            return `version: '2.1'
-
-services:
-  ${serviceName}:
-    image: ${serviceName}
-    build: .
-    ports:
-      - ${port}:${port}`;
-
-        case 'asp.net core':
-            // we don't generate compose files for .net core
-            return `version: '2.1'
-
-services:
-  ${serviceName}:
-    image: ${serviceName}
-    build: .
-    ports:
-      - ${port}:${port}`;
-
-        case 'python':
-            return `version: '2.1'
-
-services:
-  ${serviceName}:
-    image: ${serviceName}
-    build:
-      context: .
-      dockerfile: Dockerfile
-    ports:
-        - ${port}:${port}
-`;
-
-        case 'ruby':
-            return `version: '2.1'
-
-services:
-  ${serviceName}:
-    image: ${serviceName}
-    build:
-      context: .
-      dockerfile: Dockerfile
-    ports:
-        - ${port}:${port}
-`;
-
-        case 'java':
-            return `version: '2.1'
-
-services:
-  ${serviceName}:
-    image: ${serviceName}
-    build:
-      context: .
-      dockerfile: Dockerfile
-    environment:
-      JAVA_OPTS: -agentlib:jdwp=transport=dt_socket,server=y,suspend=n,address=5005,quiet=y
-    ports:
-      - ${port}:${port}
-      - 5005:5005
-    `;
-
-        default:
-            return `version: '2.1'
-
-services:
-  ${serviceName}:
-    image: ${serviceName}
-    build:
-      context: .
-      dockerfile: Dockerfile
-    ports:
-      - ${port}:${port}
-`;
+function genDockerComposeDebug(serviceName: string, platform: Platform, os: OS | undefined, port: string, packageInfo: Partial<PackageInfo>): string {
+    let generators = generatorsByPlatform.get(platform);
+    assert(generators, `Could not find docker debug compose file generator function for "${platform}"`);
+    if (generators.genDockerComposeDebug) {
+        return generators.genDockerComposeDebug(serviceName, platform, os, port, packageInfo);
     }
 }
+
+//////////////////////////////////
 
 function genDockerIgnoreFile(service: string, platformType: string, os: string, port: string): string {
     return `node_modules
@@ -603,7 +244,7 @@ async function findCSProjFile(folderPath: string): Promise<string> {
     return projectFiles[0].slice(0, -'.csproj'.length);
 }
 
-type GeneratorFunction = (serviceName: string, platform: string, os: string, port: string, packageJson?: PackageInfo) => string;
+type GeneratorFunction = (serviceName: string, platform: Platform, os: OS | undefined, port: string, packageJson?: Partial<PackageInfo>) => string;
 
 const DOCKER_FILE_TYPES: { [key: string]: GeneratorFunction } = {
     'docker-compose.yml': genDockerCompose,
