+++ configureWorkspace/configure_asp.net.core.ts
@@ -1,0 +1,66 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See LICENSE.md in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import * as assert from 'assert';
+import { PackageInfo } from './configure';
+
+export let configureAspDotNetCore = {
+  genDockerFile,
+  genDockerCompose: undefined, // We don't generate compose files for .net core
+  genDockerComposeDebug: undefined // We don't generate compose files for .net core
+};
+
+// Note: serviceName includes the path of the service relative to the generated file, e.g. 'projectFolder1/myAspNetService'
+function genDockerFile(serviceName: string, platform: string, os: string | undefined, port: string, { cmd, author, version, artifactName }: Partial<PackageInfo>): string {
+  if (os.toLowerCase() === 'windows') {
+    return `
+FROM microsoft/aspnetcore:2.0-nanoserver-1709 AS base
+WORKDIR /app
+EXPOSE ${port}
+
+FROM microsoft/aspnetcore-build:2.0-nanoserver-1709 AS build
+WORKDIR /src
+COPY ${serviceName}.csproj ${serviceName}/
+RUN dotnet restore ${serviceName}/${serviceName}.csproj
+WORKDIR /src/${serviceName}
+COPY . .
+RUN dotnet build ${serviceName}.csproj -c Release -o /app
+
+FROM build AS publish
+RUN dotnet publish ${serviceName}.csproj -c Release -o /app
+
+FROM base AS final
+WORKDIR /app
+COPY --from=publish /app .
+ENTRYPOINT ["dotnet", "${serviceName}.dll"]
+`;
+  } else {
+    // Linux
+    return `
+#Depending on the operating system of the host machines(s) that will build or run the containers, the image specified in the FROM statement may need to be changed.
+#For more information, please see http://aka.ms/containercompat
+
+FROM microsoft/aspnetcore:2.0 AS base
+WORKDIR /app
+EXPOSE ${port}
+
+FROM microsoft/aspnetcore-build:2.0 AS build
+WORKDIR /src
+COPY ${serviceName}.csproj ${serviceName}/
+RUN dotnet restore ${serviceName}/${serviceName}.csproj
+WORKDIR /src/${serviceName}
+COPY . .
+RUN dotnet build ${serviceName}.csproj -c Release -o /app
+
+FROM build AS publish
+RUN dotnet publish ${serviceName}.csproj -c Release -o /app
+
+FROM base AS final
+WORKDIR /app
+COPY --from=publish /app .
+ENTRYPOINT ["dotnet", "${serviceName}.dll"]
+`;
+  }
+}
