+++ configureWorkspace/configure_python.ts
@@ -1,0 +1,70 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See LICENSE.md in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { PackageInfo } from './configure';
+
+export let configurePython = {
+  genDockerFile,
+  genDockerCompose,
+  genDockerComposeDebug
+};
+
+// Note: serviceName includes the path of the service relative to the generated file, e.g. 'projectFolder1/myAspNetService'
+function genDockerFile(serviceName: string, platform: string, os: string | undefined, port: string, { cmd, author, version, artifactName }: Partial<PackageInfo>): string {
+  return `# Python support can be specified down to the minor or micro version
+# (e.g. 3.6 or 3.6.3).
+# OS Support also exists for jessie & stretch (slim and full).
+# See https://hub.docker.com/r/library/python/ for all supported Python
+# tags from Docker Hub.
+FROM python:alpine
+
+# If you prefer miniconda:
+#FROM continuumio/miniconda3
+
+LABEL Name=${serviceName} Version=${version}
+EXPOSE ${port}
+
+WORKDIR /app
+ADD . /app
+
+# Using pip:
+RUN python3 -m pip install -r requirements.txt
+CMD ["python3", "-m", "${serviceName}"]
+
+# Using pipenv:
+#RUN python3 -m pip install pipenv
+#RUN pipenv install --ignore-pipfile
+#CMD ["pipenv", "run", "python3", "-m", "${serviceName}"]
+
+# Using miniconda (make sure to replace 'myenv' w/ your environment name):
+#RUN conda env create -f environment.yml
+#CMD /bin/bash -c "source activate myenv && python3 -m ${serviceName}"
+`;
+}
+
+function genDockerCompose(serviceName: string, platform: string, os: string | undefined, port: string): string {
+  return `version: '2.1'
+
+services:
+  ${serviceName}:
+    image: ${serviceName}
+    build: .
+    ports:
+      - ${port}:${port}`;
+}
+
+function genDockerComposeDebug(serviceName: string, platform: string, os: string | undefined, port: string, { fullCommand: cmd }: Partial<PackageInfo>): string {
+  return `version: '2.1'
+
+services:
+  ${serviceName}:
+    image: ${serviceName}
+    build:
+      context: .
+      dockerfile: Dockerfile
+    ports:
+        - ${port}:${port}
+`;
+}
